Detect zero crossings in a signed int
We pass the sum of both channels for zero detection, which is a signed
int. So, to cast it to a signed short is a bug.

+++ timecoder.c
@@ -321,7 +321,7 @@ static void set_sample_rate(struct timecoder_t *tc, int rate)
 
 
 static int detect_zero_crossing(struct timecoder_channel_t *ch,
-                                signed short v, float alpha)
+                                signed int v, float alpha)
 {
     int swapped;
